requirements.txt
fastapi
uvicorn
python-multipart
pillow
pdfminer.six
markdownify
openai
faiss-cpu
tqdm
python-dotenv
PyMuPDF
sentence-transformers
requests
beautifulsoup4
google-generativeai